requirements.txt
accelerate==0.21.0
aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
bleach==6.1.0
blinker==1.7.0
boto3==1.34.32
botocore==1.34.32
byted-lafka-internal==1.4.16rc1
byted-remote-ikernel==0.4.8
byted-torch @ http://luban-source.byted.org/repository/scm/lab.pytorch.pytorch2_cuda122_debug_1.0.0.42.tar.gz#sha256=5f68c2728e611179e0e8d3cb5b54006463761bf298d03c22236128c496d6fb1d
byted-tracking==0.0.12
byted-wandb==0.13.72
bytedance-context==0.7.1
bytedance-metrics==0.5.1
bytedbackgrounds==0.0.6
byteddatabus==1.0.6
byteddps==0.1.2
bytedenv==0.6.2
bytedkafka==0.2.17
bytedlogger==0.15.0
bytedlogid==0.2.1
bytedmemfd==0.2
bytedmetrics==0.10.2
bytedservicediscovery==0.17.4
bytedsinfmetacenter==1.4.2
bytedtcc==1.4.4
bytedtos==1.1.16
bytedtrace==0.2.1
bytedztijwt==0.2.18
bytedztijwthelper==0.0.22
bytedztispiffe==0.0.11
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
crcmod==1.7
cryptography==38.0.4
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.12.6
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
distro-info==1.0
dnspython==2.5.0
docker-pycreds==0.4.0
ecdsa==0.18.0
einops==0.6.1
einops-exts==0.0.4
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
extension_jupyter_kernel @ file:///opt/tiger/mlx_deploy/jupyter/1.0.0.177/extension_jupyter_kernel-0.1.0.tar.gz#sha256=aaff04052cda68527d7ea0c18a797afff92d7c9e5b716e09653ac81a5b740384
fastapi==0.110.0
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.9.0
findspark==2.0.1
flash-attn==2.5.7
Flask==3.0.1
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.3.0
iotop==0.6
ipaddress==1.0.23
ipykernel==6.29.4
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-kernel-gateway==2.5.2
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_fileid==0.9.1
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
merlin_kernel @ file:///opt/tiger/mlx_deploy/python_kernel/1.0.0.395
mistune==3.0.2
mlx-python-sdk @ http://luban-source.byted.org/repository/scm/data.reckon.mlx_python_sdk_1.0.0.207.tar.gz#sha256=4134435ed17cf7acba03f20a6e06cda9e0336c675bcd6cd7b5a48c8ad6b15e6b
mpmath==1.2.1
msgpack==1.0.7
multidict==6.0.5
namesgenerator==0.3
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==2.6.3
ninja==1.11.1.1
notebook==6.5.6
notebook_shim==0.2.4
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
opencv-python==4.9.0.80
openpyxl==3.1.2
orjson==3.9.15
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
peft==0.9.0
pexpect==4.8.0
Pillow==9.5.0
pipdeptree==2.16.2
platformdirs==4.2.0
ply==3.11
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyasn1==0.5.1
pycparser==2.21
pycurl==7.43.0.6
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.38.0
PyJWT==1.7.1
pynvml==11.5.0
pyOpenSSL==22.1.0
pyparsing==3.1.2
python-apt==2.2.1
python-dateutil==2.8.2
python-etcd==0.4.5
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-snappy==0.6.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
reactivex==4.0.4
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.3
s3transfer==0.10.0
safetensors==0.4.2
schedule==1.2.1
scikit-learn==1.2.2
scipy==1.12.0
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.44.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.36.3
svgwrite==1.4.3
sympy==1.11.1
terminado==0.18.1
threadpoolctl==3.3.0
thrift==0.16.0
thriftpy2==0.4.16
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.2
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.28.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
y-py==0.6.2
yarl==1.9.4
ypy-websocket==0.8.4
zipp==3.17.0
zstandard==0.22.0
